moving page objects to folders, refactor entity list
