+++ ATlearning/ATframework3demo/PageObjects/CreateForms/LessonCreateForm.cs
@@ -2,7 +2,7 @@
 using atFrameWork2.SeleniumFramework;
 using ATframework3demo.TestEntities;
 
-namespace ATframework3demo.PageObjects
+namespace ATframework3demo.PageObjects.CreateForms
 {
     public class LessonCreateForm
     {
