+++ ATlearning/ATframework3demo/PageObjects/EditForms/GroupEditForm.cs
@@ -2,7 +2,7 @@
 using atFrameWork2.SeleniumFramework;
 using ATframework3demo.TestEntities;
 
-namespace ATframework3demo.PageObjects
+namespace ATframework3demo.PageObjects.EditForms
 {
     public class GroupEditForm
     {
