+++ ATlearning/ATframework3demo/PageObjects/Lists/SubjectList.cs
@@ -1,7 +1,9 @@
 ﻿using atFrameWork2.PageObjects;
 using atFrameWork2.SeleniumFramework;
+using ATframework3demo.PageObjects.CreateForms;
+using ATframework3demo.TestEntities;
 
-namespace ATframework3demo.PageObjects
+namespace ATframework3demo.PageObjects.Lists
 {
     public class SubjectList : EntityList
     {
