+++ ATlearning/ATframework3demo/PageObjects/Lists/UserList.cs
@@ -1,8 +1,10 @@
 ﻿using atFrameWork2.PageObjects;
 using atFrameWork2.SeleniumFramework;
+using ATframework3demo.PageObjects.CreateForms;
+using ATframework3demo.PageObjects.EditForms;
 using ATframework3demo.TestEntities;
 
-namespace ATframework3demo.PageObjects
+namespace ATframework3demo.PageObjects.Lists
 {
     public class UserList : EntityList
     {
